PP9-9325 Angular SDK - update sample application
