--- samples/picturepark-sdk-v1-sample-angular7/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "picturepark-sdk-v1-sample-angular6",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "oidc-client": "^1.4.1",
-    "angular2-virtual-scroll": "^0.3.1",
-    "@angular/animations": "^7.1.3",
-    "@angular/common": "^7.1.3",
-    "@angular/compiler": "^7.1.3",
-    "@angular/core": "^7.1.3",
-    "@angular/forms": "^7.1.3",
-    "@angular/http": "^7.1.3",
-    "@angular/platform-browser": "^7.1.3",
-    "@angular/platform-browser-dynamic": "^7.1.3",
-    "@angular/router": "^7.1.3",
-    "@picturepark/sdk-v1-angular": "^0.3.0",
-    "core-js": "^2.5.4",
-    "rxjs": "^6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.26"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.1",
-    "@angular/cli": "~7.2.1",
-    "@angular/compiler-cli": "^7.1.3",
-    "@angular/language-service": "^7.1.3",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~3.2.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^1.3.0",
-    "protractor": "^5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.1.6"
-  }
-}
