+++ samples/picturepark-sdk-v1-sample-angular8/README.md
@@ -1,6 +1,6 @@
-# PictureparkSdkV1SampleAngular7
+# PictureparkSdkV1SampleAngular8
 
-This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 6.0.5.
+This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 8.3.4.
 
 ## Development server
