+++ samples/picturepark-sdk-v1-sample-angular8/angular.json
@@ -3,27 +3,32 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "picturepark-sdk-v1-sample-angular6": {
+    "picturepark-sdk-v1-sample-angular8": {
+      "projectType": "application",
+      "schematics": {
+        "@schematics/angular:component": {
+          "style": "scss"
+        }
+      },
       "root": "",
       "sourceRoot": "src",
-      "projectType": "application",
       "prefix": "app",
-      "schematics": {},
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/picturepark-sdk-v1-sample-angular6",
+            "outputPath": "dist/picturepark-sdk-v1-sample-angular8",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
-            "tsConfig": "src/tsconfig.app.json",
+            "tsConfig": "tsconfig.app.json",
+            "aot": false,
             "assets": [
               "src/favicon.ico",
               "src/assets"
             ],
             "styles": [
-              "src/styles.css"
+              "src/styles.scss"
             ],
             "scripts": []
           },
@@ -43,25 +48,37 @@
               "aot": true,
               "extractLicenses": true,
               "vendorChunk": false,
-              "buildOptimizer": true
+              "buildOptimizer": true,
+              "budgets": [
+                {
+                  "type": "initial",
+                  "maximumWarning": "2mb",
+                  "maximumError": "5mb"
+                },
+                {
+                  "type": "anyComponentStyle",
+                  "maximumWarning": "6kb",
+                  "maximumError": "10kb"
+                }
+              ]
             }
           }
         },
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "picturepark-sdk-v1-sample-angular6:build"
+            "browserTarget": "picturepark-sdk-v1-sample-angular8:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "picturepark-sdk-v1-sample-angular6:build:production"
+              "browserTarget": "picturepark-sdk-v1-sample-angular8:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "picturepark-sdk-v1-sample-angular6:build"
+            "browserTarget": "picturepark-sdk-v1-sample-angular8:build"
           }
         },
         "test": {
@@ -69,54 +86,44 @@
           "options": {
             "main": "src/test.ts",
             "polyfills": "src/polyfills.ts",
-            "tsConfig": "src/tsconfig.spec.json",
-            "karmaConfig": "src/karma.conf.js",
-            "styles": [
-              "src/styles.css"
-            ],
-            "scripts": [],
+            "tsConfig": "tsconfig.spec.json",
+            "karmaConfig": "karma.conf.js",
             "assets": [
               "src/favicon.ico",
               "src/assets"
-            ]
+            ],
+            "styles": [
+              "src/styles.scss"
+            ],
+            "scripts": []
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "src/tsconfig.app.json",
-              "src/tsconfig.spec.json"
+              "tsconfig.app.json",
+              "tsconfig.spec.json",
+              "e2e/tsconfig.json"
             ],
             "exclude": [
               "**/node_modules/**"
             ]
           }
-        }
-      }
-    },
-    "picturepark-sdk-v1-sample-angular6-e2e": {
-      "root": "e2e/",
-      "projectType": "application",
-      "architect": {
+        },
         "e2e": {
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "picturepark-sdk-v1-sample-angular6:serve"
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": "e2e/tsconfig.e2e.json",
-            "exclude": [
-              "**/node_modules/**"
-            ]
+            "devServerTarget": "picturepark-sdk-v1-sample-angular8:serve"
+          },
+          "configurations": {
+            "production": {
+              "devServerTarget": "picturepark-sdk-v1-sample-angular8:serve:production"
+            }
           }
         }
       }
-    }
-  },
-  "defaultProject": "picturepark-sdk-v1-sample-angular6"
+    }},
+  "defaultProject": "picturepark-sdk-v1-sample-angular8"
 }
