+++ samples/picturepark-sdk-v1-sample-angular8/src/app/app.component.ts
@@ -4,7 +4,7 @@ import { ShareService } from '@picturepark/sdk-v1-angular';
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
   output: string;
