+++ samples/picturepark-sdk-v1-sample-angular8/src/app/app.module.ts
