+++ samples/picturepark-sdk-v1-sample-angular8/src/environments/environment.prod.ts
