+++ samples/picturepark-sdk-v1-sample-angular8/src/index.html
@@ -2,9 +2,8 @@
 <html lang="en">
 <head>
   <meta charset="utf-8">
-  <title>PictureparkSdkV1SampleAngular6</title>
+  <title>PictureparkSdkV1SampleAngular8</title>
   <base href="/">
-
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
