+++ samples/picturepark-sdk-v1-sample-angular8/src/test.ts
